pyproject.toml
[tool.poetry]
name = "ghostpost_backend"
version = "0.1.0"
description = "Back End of Ghost Post Rest Framework"
authors = ["Cheria Artis"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.12.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
